% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{add_pref}
\alias{add_pref}
\title{Add Prefix}
\usage{
add_pref(pref, x)
}
\arguments{
\item{pref}{character; prefix to add}

\item{x}{character; string to change}
}
\value{
character
}
\description{
Add Prefix
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
add_pref('census_', x)
}
\concept{prefix}
