% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix.R
\name{add_suff}
\alias{add_suff}
\title{Add Suffix}
\usage{
add_suff(suff, x)
}
\arguments{
\item{suff}{character; suffix to add}

\item{x}{character; string to change}
}
\value{
character
}
\description{
Add Suffix
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
add_suff('_cen', x)
}
\concept{suffix}
