% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_names}
\alias{clean_names}
\title{Clean data frame column names}
\usage{
clean_names(.data, unique = FALSE)
}
\arguments{
\item{.data}{a data frame}

\item{unique}{logical; if \code{TRUE}, duplicate names after cleaning are made
unique by appending \verb{_1}, \verb{_2}, etc. Default is \code{FALSE}.}
}
\value{
a data frame with cleaned column names
}
\description{
Renames all columns of a data frame using \code{\link[=str_clean_names]{str_clean_names()}}.
}
\examples{
tb <- tibble::tibble(popTotal = 1, `Pop Total` = 2, pop.total = 3)
clean_names(tb)
clean_names(tb, unique = TRUE)
}
\concept{clean}
