% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_names}
\alias{compare_names}
\title{Compare the Names of Two Objects}
\usage{
compare_names(x, y)
}
\arguments{
\item{x}{first object}

\item{y}{second object}
}
\value{
character vector of differences, invisibly
}
\description{
Compare the Names of Two Objects
}
\examples{
s <- tibble::tibble(a = 1, b = 2, d = 3)
t <- tibble::tibble(a = 1, d = 3, c = 2)
compare_names(s, t)
}
\concept{compare}
