% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_phrase}
\alias{list_phrase}
\title{List Phrases}
\usage{
list_phrase(tb, loc = 2)
}
\arguments{
\item{tb}{tibble; data to list prefixes in}

\item{loc}{number of location to list. For example \verb{1_2_3_4} with \code{loc = 3} returns 3.}
}
\value{
character
}
\description{
List Phrases
}
\examples{
tb <- tibble::tibble(pop = 10, pop_2020_est = 9, pop_white_2020 = 8, pop_black_2020 = 2)
list_phrase(tb)
}
\concept{list}
