% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_suff}
\alias{list_suff}
\title{List Suffixes}
\usage{
list_suff(tb)
}
\arguments{
\item{tb}{tibble; data to list prefixes in}
}
\value{
character
}
\description{
List Suffixes
}
\examples{
tb <- tibble::tibble(pop = 10, pop_2020_est = 9, pop_white_2020 = 8, pop_black_2020 = 2)
list_suff(tb)
}
\concept{list}
