% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{rem_pref}
\alias{rem_pref}
\title{Remove Prefix}
\usage{
rem_pref(x, pref)
}
\arguments{
\item{x}{character; string to change}

\item{pref}{character; prefix to remove}
}
\value{
character
}
\description{
Remove Prefix
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
rem_pref(x, 'pop_')
}
\concept{prefix}
