% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix.R
\name{rem_suff}
\alias{rem_suff}
\title{Remove Suffix}
\usage{
rem_suff(x, suff)
}
\arguments{
\item{x}{character; string to change}

\item{suff}{character; suffix to remove}
}
\value{
character
}
\description{
Remove Suffix
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
rem_suff(x, '_2020')
}
\concept{suffix}
