% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase.R
\name{repl_phrase}
\alias{repl_phrase}
\title{Replace Phrase}
\usage{
repl_phrase(phrase, repl, x)
}
\arguments{
\item{phrase}{character; phrase to replace}

\item{repl}{character; phrase to replace with}

\item{x}{character; string to change}
}
\value{
character
}
\description{
Replace Phrase
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
repl_phrase('_2020', '_20', x)
}
\concept{phrase}
