% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{repl_pref}
\alias{repl_pref}
\title{Replace Prefix}
\usage{
repl_pref(pref, repl, x)
}
\arguments{
\item{pref}{character; prefix to replace}

\item{repl}{character; prefix to replace with}

\item{x}{character; string to change}
}
\value{
character
}
\description{
Replace Prefix
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
repl_pref('pop_', 'p_', x)
}
\concept{prefix}
