% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix.R
\name{repl_suff}
\alias{repl_suff}
\title{Replace Suffix}
\usage{
repl_suff(suff, repl, x)
}
\arguments{
\item{suff}{character; suffix to replace}

\item{repl}{character; suffix to replace with}

\item{name}{character; string to change}
}
\value{
character
}
\description{
Replace Suffix
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
repl_suff('_2020', '_20', x)
}
\concept{suffix}
