% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sd}
\alias{sd}
\title{South Dakota Election and Demographic Data}
\usage{
data("sd")
}
\description{
This data set contains demographic and election information for South Dakota
}
\examples{
data(sd)
}
\references{
Voting and Election Science Team, 2020, "2020 Precinct-Level Election Results",
https://doi.org/10.7910/DVN/K7760H, Harvard Dataverse, V23

Voting and Election Science Team, 2018, "2016 Precinct-Level Election Results",
https://doi.org/10.7910/DVN/NH5S2I, Harvard Dataverse, V71

Voting and Election Science Team, 2019, "2018 Precinct-Level Election Results",
https://doi.org/10.7910/DVN/UBKYRU, Harvard Dataverse, V48

Kenny & McCartan (2021, Aug. 10). ALARM Project: 2020 Redistricting Data Files.
Retrieved from \url{https://github.com/alarm-redist/census-2020/}
}
\concept{data}
