% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort_pref}
\alias{sort_pref}
\title{Sort by Prefix}
\usage{
sort_pref(x)
}
\arguments{
\item{x}{character; strings to sort}
}
\value{
character
}
\description{
Sort by Prefix
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
sort_pref(x)
}
\concept{sort}
