% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort_suff}
\alias{sort_suff}
\title{Sort by Suffix}
\usage{
sort_suff(x)
}
\arguments{
\item{x}{character; strings to sort}
}
\value{
character
}
\description{
Sort by Suffix
}
\examples{
x <- c('pop_2020_est', 'pop_white_2020', 'pop_black_2020', 'pop_white_2021')
sort_suff(x)
}
\concept{sort}
