% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase.R
\name{rem_phrase}
\alias{rem_phrase}
\title{Remove Phrase}
\usage{
rem_phrase(x, phrase)
}
\arguments{
\item{x}{character; string to change}

\item{phrase}{character; phrase to remove}
}
\value{
character
}
\description{
Remove Phrase
}
\examples{
x <- c('pop', 'pop_2020_est', 'pop_white_2020', 'pop_black_2020')
rem_phrase(x, '_2020')
}
\concept{phrase}
